Rename momoapi to momoapi-ruby
