+++ lib/momoapi-ruby/errors.rb
