+++ lib/momoapi-ruby/version.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 module Momoapi
-  module Ruby
-    VERSION = '0.1.0'
-  end
+  VERSION = '0.1.0'
 end
