--- lib/momoapi/ruby.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require 'momoapi/ruby/version'
-
-module Momoapi
-  module Ruby
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
